fix: move snackbars, and use a separate viewmodel
