+++ composeApp/src/androidMain/kotlin/paige/navic/di/PlatformModule.android.kt
@@ -53,7 +53,8 @@ actual val platformModule = module {
 			downloadManager = get(),
 			connectivityManager = get(),
 			sessionManager = get(),
-			preferenceManager = get()
+			preferenceManager = get(),
+			snackBarViewModel = get()
 		)
 	}
