+++ composeApp/src/commonMain/kotlin/paige/navic/domain/models/snackbars/PlayerEvent.kt
@@ -1,0 +1,8 @@
+package paige.navic.domain.models.snackbars
+
+import org.jetbrains.compose.resources.StringResource
+
+data class PlayerEvent(
+	val resource: StringResource,
+	val args: List<Any> = emptyList()
+)
