+++ composeApp/src/commonMain/kotlin/paige/navic/shared/MediaPlayer.kt
@@ -3,20 +3,13 @@ package paige.navic.shared
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.launch
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.decodeFromJsonElement
-import navic.composeapp.generated.resources.Res
-import navic.composeapp.generated.resources.notice_added_to_queue
-import navic.composeapp.generated.resources.notice_play_next
-import org.jetbrains.compose.resources.StringResource
 import paige.navic.domain.models.DomainRadio
 import paige.navic.domain.models.DomainSong
 import paige.navic.domain.models.DomainSongCollection
@@ -27,32 +20,15 @@ import paige.navic.ui.core.PlayerUiState
 import paige.navic.util.core.Logger
 import kotlin.time.Duration.Companion.seconds
 
-data class PlayerEvent(
-	val resource: StringResource,
-	val args: List<Any> = emptyList()
-)
-
 abstract class MediaPlayerViewModel(
 	private val stateRepository: PlayerStateRepository,
 	protected val connectivityManager: ConnectivityManager,
-	protected val downloadManager: DownloadManager
+	protected val downloadManager: DownloadManager,
 ) : ViewModel() {
 
 	@Suppress("PropertyName")
 	protected val _uiState = MutableStateFlow(PlayerUiState())
 	val uiState: StateFlow<PlayerUiState> = _uiState.asStateFlow()
-
-	private val _events = MutableSharedFlow<PlayerEvent>()
-	val events: SharedFlow<PlayerEvent> = _events.asSharedFlow()
-
-	fun notify(resource: StringResource, vararg args: Any) {
-		viewModelScope.launch {
-			_events.emit(PlayerEvent(resource, args.toList()))
-		}
-	}
-
-	protected fun notifyAddedToQueue() = notify(Res.string.notice_added_to_queue)
-	protected fun notifyPlayNext() = notify(Res.string.notice_play_next)
 
 	protected fun isAvailable(songId: String): Boolean {
 		val isOnline = connectivityManager.isOnline.value
