+++ composeApp/src/commonMain/kotlin/paige/navic/ui/components/snackbars/ErrorSnackbar.kt
@@ -1,4 +1,4 @@
-package paige.navic.ui.components.common
+package paige.navic.ui.components.snackbars
 
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarResult
@@ -16,6 +16,8 @@ import navic.composeapp.generated.resources.info_error_show
 import org.jetbrains.compose.resources.getString
 import org.jetbrains.compose.resources.stringResource
 import paige.navic.LocalSnackbarState
+import paige.navic.ui.components.common.ErrorCodeBlock
+import paige.navic.ui.components.common.FormButton
 import paige.navic.util.core.Logger
 import paige.navic.ui.components.dialogs.FormDialog
 
@@ -51,14 +53,14 @@ fun ErrorSnackbar(
 			onClearError()
 		},
 		buttons = {
-			FormButton(onClick = {
-				visible = false
-				onClearError()
-			}) {
-				Text(stringResource(Res.string.action_ok))
-			}
+            FormButton(onClick = {
+                visible = false
+                onClearError()
+            }) {
+                Text(stringResource(Res.string.action_ok))
+            }
 		}
 	) {
-		ErrorCodeBlock(error)
+        ErrorCodeBlock(error)
 	}
 }
