+++ composeApp/src/commonMain/kotlin/paige/navic/ui/screens/radio/RadioListScreen.kt
@@ -45,7 +45,7 @@ import paige.navic.domain.models.settings.BottomBarCollapseMode
 import paige.navic.icons.Icons
 import paige.navic.icons.outlined.Add
 import paige.navic.shared.MediaPlayerViewModel
-import paige.navic.ui.components.common.ErrorSnackbar
+import paige.navic.ui.components.snackbars.ErrorSnackbar
 import paige.navic.ui.components.layouts.ArtGrid
 import paige.navic.ui.components.layouts.NestedTopBar
 import paige.navic.ui.components.layouts.PullToRefreshBox
