+++ composeApp/src/iosMain/kotlin/paige/navic/di/PlatformModule.ios.kt
@@ -60,7 +60,8 @@ actual val platformModule = module {
 			connectivityManager = get(),
 			syncManager = get(),
 			sessionManager = get(),
-			preferenceManager = get()
+			preferenceManager = get(),
+			snackBarViewModel = get()
 		)
 	}
